split g2 into 3 files
